pyproject.toml
[tool.poetry]
name = "quantbot"
version = "2.0.0"
description = "QuantBot Phase 2 — Factor Component Engine"
authors = ["Devon Lopez"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# Numerical + Data
numpy = "^1.26"
pandas = "^2.2"
scipy = "^1.12"
scikit-learn = "^1.4"
statsmodels = "^0.14"

# Financial Data
yfinance = "^0.2"

# NLP / Sentiment
transformers = "^4.40"
torch = "^2.3"
sentencepiece = "^0.2"

# Async Data Collection
aiohttp = "^3.9"
httpx = "^0.27"
tenacity = "^8.3"

# Storage
pyarrow = "^15.0"
duckdb = "^0.10"
sqlalchemy = "^2.0"

# Config + Validation
pydantic = "^2.7"
pandera = "^0.19"
pyyaml = "^6.0"

# Caching (optional)
redis = {version = "^5.0", optional = true}

[tool.poetry.extras]
redis = ["redis"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.23"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
